requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
passlib==1.7.4
pydantic==1.10.8
uvicorn==0.22.0
mysql-connector-python==8.0.33
google-cloud-storage==2.9.0
PyJWT==2.7.0
SQLAlchemy==2.0.15
requests==2.31.0
python-multipart==0.0.6
